python/requirements.txt
# --------- base-tool --------- #
sphinx==8.0.2
sphinx-rtd-theme==0.5.1
pytest==8.3.3
pytest-cov==5.0.0
pyrootutils==1.0.4
pre-commit==3.8.0
twine==5.1.1
wheel>=0.41.3
coverage==7.6.1
furo==2024.8.6
myst-parser==4.0.0
pytest-mock==3.14.0
pendulum==3.0.0

# --------- date --------- #
openpyxl==3.1.5
attrs>=23.1.0,<24

# --------- path --------- #
pathlib==1.0.1
watchdog
# --------- web --------- #
streamlit==1.39.0

# --------- cli --------- #
# No entries

# --------- log --------- #
loguru==0.7.2
prettytable==3.11.0

# --------- ai-tools --------- #
langchain-openai==0.1.25
langchain-community==0.2.17
langchain_cohere==0.1.9
langchain_postgres==0.0.12
openai
duckduckgo-search==6.2.13
whisper==1.1.10
arxiv==2.1.3

# --------- server --------- #
fastapi==0.115.0
uvicorn==0.31.0

# --------- db --------- #
psycopg==3.2.3
psycopg-binary==3.2.3
langchain-chroma==0.1.4
chromadb==0.4.24
# faiss-cpu is commented out

# --------- deploy --------- #
# ansible is commented out

# --------- data --------- #
pydantic==2.9.2
numpy==1.26.4
pandas==2.2.3
pypdf==4.3.1
pyarrow==17.0.0
unstructured==0.15.13
python-docx==1.1.2
python-magic==0.4.27
python-pptx==1.0.2
docx2txt==0.8
markdown==3.7

# --------- graph database --------- #
neo4j==5.25.0
# neomodel[extras] is commented out

# --------- yaml --------- #
pyyaml==6.0.2

# --------- agent --------- #
crewai==0.67.1
crewai-tools==0.12.1
dspy-ai==2.5.3
mem0ai==0.1.17

# --------- distributed --------- #
dora-rs

# --------- stock --------- #
yfinance==0.2.44

# --------- embedding --------- #
sentence-transformers==3.1.1
transformers==4.45.1

# --------- ultralytics --------- #
ultralytics==8.3.5


# --------- web sider --------- #
scrapegraphai
undetected_chromedriver
selenium